# backend/driver/openapi.yaml
# Generated with protoc-gen-openapi
# https://github.com/google/gnostic/tree/master/cmd/protoc-gen-openapi

openapi: 3.0.3
info:
    title: Driver API
    version: 0.0.1
paths:
    /driver/get-verify-code/{telephone}:
        get:
            tags:
                - Driver
            operationId: Driver_GetVerifyCode
            parameters:
                - name: telephone
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/GetVerifyCodeResp'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
    /driver/login:
        post:
            tags:
                - Driver
            operationId: Driver_Login
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/LoginReq'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/LoginResp'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
    /driver/logout:
        get:
            tags:
                - Driver
            operationId: Driver_Logout
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/LogoutResp'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
components:
    schemas:
        GetVerifyCodeResp:
            type: object
            properties:
                code:
                    type: integer
                    format: int32
                message:
                    type: string
                verifyCode:
                    type: string
                verifyCodeTime:
                    type: integer
                    description: 生成时间
                    format: int64
                verifyCodeLife:
                    type: integer
                    description: 存活时间
                    format: int32
        GoogleProtobufAny:
            type: object
            properties:
                '@type':
                    type: string
                    description: The type of the serialized message.
            additionalProperties: true
            description: Contains an arbitrary serialized message along with a @type that describes the type of the serialized message.
        LoginReq:
            type: object
            properties:
                telephone:
                    type: string
                verifyCode:
                    type: string
        LoginResp:
            type: object
            properties:
                code:
                    type: integer
                    format: int32
                message:
                    type: string
                token:
                    type: string
                tokenCreateAt:
                    type: integer
                    description: 生成时间
                    format: int64
                tokenLife:
                    type: integer
                    description: 存活时间
                    format: int64
        LogoutResp:
            type: object
            properties:
                code:
                    type: integer
                    format: int32
                message:
                    type: string
        Status:
            type: object
            properties:
                code:
                    type: integer
                    description: The status code, which should be an enum value of [google.rpc.Code][google.rpc.Code].
                    format: int32
                message:
                    type: string
                    description: A developer-facing error message, which should be in English. Any user-facing error message should be localized and sent in the [google.rpc.Status.details][google.rpc.Status.details] field, or localized by the client.
                details:
                    type: array
                    items:
                        $ref: '#/components/schemas/GoogleProtobufAny'
                    description: A list of messages that carry the error details.  There is a common set of message types for APIs to use.
            description: 'The `Status` type defines a logical error model that is suitable for different programming environments, including REST APIs and RPC APIs. It is used by [gRPC](https://github.com/grpc). Each `Status` message contains three pieces of data: error code, error message, and error details. You can find out more about this error model and how to work with it in the [API Design Guide](https://cloud.google.com/apis/design/errors).'
tags:
    - name: Driver